Implemented revealFog and highLightSquares with some issues:
Both take in a list of squares and do them all at once. But
1, tweens don't work on tiles (their alpha changes, but it
appears no one is checking their alpha so they just look the
same until they are removed). And 2, given the way we've set
this all up, I can't figure out a way to not have the tweens
overwrite each other, causing only the last one to happen
(as in, only the last tile in the list highlights or is revealed)