Fix phaser loading issues
The reason why phaser only loads when you refresh the page is because
of turbolinks which is a rails 4 default. I changed the event to listen
to a turbolinks event instead of window.onload. I also fixed the game
loading on random pages with a manual check for the parent id.